requirements.txt
adal==1.2.7
advocate==1.0.0
alembic==1.12.0
aniso8601==8.0.0
annotated-types==0.6.0
anyio==4.0.0
appdirs==1.4.4
arrow==1.3.0
asn1crypto==1.5.1
asttokens==2.4.0
async-timeout==4.0.3
atsd-client==3.0.5
attrs==23.1.0
Authlib==0.15.5
azure-kusto-data==0.0.35
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bitarray==2.8.2
blessed==1.20.0
blinker==1.6.2
blis==0.7.11
boto==2.49.0
boto3==1.28.8
botocore==1.31.8
build==0.10.0
CacheControl==0.13.1
cachetools==5.3.1
cassandra-driver==3.21.0
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.0
ciso8601==2.3.0
cleo==2.0.1
click==8.1.3
cloudpathlib==0.15.1
cmem-cmempy==21.2.3
confection==0.1.3
contexttimer==0.3.3
coverage==7.2.7
crashtest==0.4.1
croniter==1.4.1
cryptography==41.0.4
cymem==2.0.8
databend-py==0.4.6
databend-sqlalchemy==0.2.4
DateTime==5.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
disposable-email-domains==0.0.95
distlib==0.3.7
distro==1.8.0
dnspython==2.4.2
docker==6.1.3
dulwich==0.21.6
elementpath==4.1.5
environs==9.5.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
filelock==3.12.4
Flask==2.3.2
Flask-Limiter==3.3.1
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==2.5.1
flask-talisman==0.7.0
Flask-WTF==1.1.1
freezegun==1.2.1
funcy==1.13
future==0.18.3
futures==3.0.5
geomet==0.1.2
gevent==23.9.1
gitdb==4.0.10
GitPython==3.1.37
google==3.0.0
google-api-python-client==1.7.11
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
grafanalib==0.7.0
greenlet==2.0.2
grpcio==1.59.0
gspread==5.11.2
gunicorn==20.0.4
h11==0.14.0
httpcore==0.16.3
httplib2==0.19.0
httpx==0.23.3
humanfriendly==10.0
identify==2.5.30
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
impyla==0.16.0
inflection==0.3.1
influxdb==5.2.3
iniconfig==2.0.0
installer==0.7.0
ipython==8.16.1
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
joblib==1.3.2
jsonschema==3.1.1
jwcrypto==1.5.0
keyring==24.2.0
langcodes==3.3.0
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
limits==3.6.0
make==0.1.6.post2
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.20.1
matplotlib-inline==0.1.6
maxminddb==2.4.0
maxminddb-geolite2==2018.703
mdurl==0.1.2
memsql==3.2.0
minio==7.1.17
mock==5.0.2
more-itertools==10.1.0
msgpack==1.0.7
murmurhash==1.0.10
mysql-connector==2.2.9
mysqlclient==2.1.1
ndg-httpsclient==0.5.1
netifaces==0.11.0
nodeenv==1.8.0
npm==0.1.1
numpy==1.26.0
nvm==1.0.21
nzalchemy==11.0.2
nzpy==1.15
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.0.7
optional-django==0.1.0
oracledb==1.4.0
ordered-set==4.1.0
oscrypto==1.3.0
packaging==23.2
pandas==1.3.4
paramiko==3.3.1
parsedatetime==2.4
parso==0.8.3
passlib==1.7.3
pathy==0.10.2
pexpect==4.8.0
phoenixdb==0.7
pickleshare==0.7.5
pinotdb==5.1.0
pkginfo==1.9.6
platformdirs==3.8.1
pluggy==1.3.0
ply==3.11
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
postgres==4.0
pre-commit==3.3.3
preshed==3.0.9
prompt-toolkit==3.0.39
protobuf==3.18.3
psutil==5.9.5
psycopg2-binary==2.9.6
psycopg2-pool==1.1
ptpython==3.0.23
ptvsd==4.3.2
ptyprocess==0.7.0
pure-eval==0.2.2
pure-sasl==0.6.2
pwgen==0.8.2.post0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyathena==1.11.5
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pydgraph==2.0.2
pydruid==0.5.7
pyexasol==0.12.0
Pygments==2.16.1
pygridgain==1.4.0
PyHive==0.6.1
pyignite==0.6.1
PyJWT==2.4.0
pymongo==4.3.3
pymssql==2.2.8
PyNaCl==1.5.0
pyodbc==4.0.28
pyOpenSSL==23.2.0
pyparsing==2.4.7
pypd==1.1.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pysaml2==7.3.1
pystache==0.6.0
pytest==7.4.0
pytest-cov==4.1.0
python-arango==6.1.0
python-dateutil==2.8.0
python-dotenv==0.19.2
python-rapidjson==1.1
pytz==2023.3.post1
PyYAML==6.0.1
qds-sdk==1.16.1
rapidfuzz==2.15.2
rdflib==6.3.2
-e git+https://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@github.com/Merdav212/iFlab.git@7182ea30379dcd5dce4fb4980618d472a0b23976#egg=redash&subdirectory=redash
redash-python==0.3.27
redis==4.6.0
redis-cli==1.0.1
redis-tools==0.0.2
regex==2023.8.8
requests==2.31.0
requests-aws-sign==0.1.5
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
RestrictedPython==6.2
rfc3986==1.5.0
rich==13.6.0
rq==1.9.0
rq-scheduler==0.11.0
rsa==4.9
ruff==0.0.289
s3transfer==0.6.2
sasl==0.3.1
scikit-learn==1.3.1
scipy==1.11.3
scramp==1.1.0
SecretStorage==3.3.3
semver==2.8.1
sentry-sdk==1.28.1
shellingham==1.5.3
simple-salesforce==0.74.3
simplejson==3.16.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
snowflake-connector-python==3.1.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.3.24
SQLAlchemy-Searchable==1.2.0
SQLAlchemy-Utils==0.34.2
sqlparse==0.4.4
srsly==2.4.8
sshtunnel==0.1.5
stack-data==0.6.3
statsd==3.3.0
supervisor==4.1.0
supervisor-checks==0.8.1
td-client==1.0.0
tenacity==8.2.3
thinc==8.2.1
threadpoolctl==3.2.0
thrift==0.16.0
thrift-sasl==0.4.3
thriftpy2==0.4.17
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.66.1
traitlets==5.11.2
trino==0.327.0
trove-classifiers==2023.10.18
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==2.1
ua-parser==0.18.0
uritemplate==3.0.1
urllib3==1.26.18
user-agents==2.0
uWSGI==2.0.22
uwsgitop==0.11
validators==0.22.0
vertica-python==1.1.1
virtualenv==20.24.1
wadllib==1.3.6
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.8
weasel==0.3.2
websocket-client==1.6.3
Werkzeug==2.3.6
wrapt==1.15.0
Wraptor==0.7.0
WTForms==2.2.1
xlrd==2.0.1
XlsxWriter==1.2.2
xmlschema==2.5.0
zipp==3.17.0
zope.event==5.0
zope.interface==6.0
